requirements.txt
requests
google-generativeai
python-dotenv
flask
jinja2
